+++ R/problems.R
@@ -46,7 +46,26 @@ warn_problems <- function(x, name = "input") {
   if (n == 0)
     return(x)
 
-  warning(n, " problems parsing ", name, ". ",
-    "See problems(...) for more details.", call. = FALSE)
+  probs <- attr(x, "problems")
+  many_problems <- nrow(probs) > 5
+
+  probs_f <- format(head(probs, 5), justify = "left")
+  probs_f[probs_f == "NA"] <- "--"
+  probs_f <- rbind(names(probs), probs_f)
+  probs_f <- lapply(probs_f, format, justify = "right")
+
+  if (many_problems) {
+    width <- vapply(probs_f, function(x) max(nchar(x)), integer(1))
+    dots <- vapply(width, function(i) paste(rep(".", i), collapse = ""),
+      FUN.VALUE = character(1))
+
+    probs_f <- Map(c, probs_f, dots)
+  }
+
+  probs_f <- do.call(paste, c(probs_f, list(sep = " ", collapse = "\n")))
+  warning(n, " parsing failure", if (n > 1) "s", ".\n", probs_f,
+    if (many_problems) "\n.See problems(...) for more details.",
+    call. = FALSE, immediate. = TRUE, noBreaks. = TRUE)
+
   x
 }